backend/requirements.txt
# FastAPI Core
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Database
sqlalchemy==2.0.41
alembic==1.16.4
asyncpg==0.30.0
redis==6.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Background Tasks
celery==5.4.0
kombu==5.5.0

# HTTP & WebSockets
requests==2.32.3
aiohttp==3.12.1
websockets==14.1


# Utilities
python-dotenv==1.0.1
python-slugify==8.0.6
pytz==2025.1
phonenumbers==8.13.56
pydantic[email]==2.11.7

# AI
openai==1.97.0

# Testing
pytest==8.4.0
pytest-asyncio==0.25.0
httpx==0.28.1
pytest-mock==3.15.0
factory-boy==3.4.0
pytest-cov==6.1.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis[asyncio]>=5.0.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic-settings>=2.1.0